Both premium calculation csv's exported.
These contain key information of each premium calculation such as profit, premium, and all the relevant indicators.

The first 18 correspond to before intervention, and csv's numbered 19-108 are the after intervention. Note that I did more policyholder types for after intervention due to the introduction of lapsed smoker and the second mortality set.